% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colAnyMissings.R
\docType{methods}
\name{colAnyMissings}
\alias{colAnyMissings}
\alias{colAnyMissings,DelayedMatrix-method}
\title{Checks if there are any missing values in an object or not}
\usage{
colAnyMissings(x, rows = NULL, cols = NULL, ...)

\S4method{colAnyMissings}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  force_block_processing = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link[base]{vector}}, a \code{\link[base]{list}}, a
\code{\link[base]{matrix}}, a \code{\link[base]{data.frame}}, or
\code{\link[base]{NULL}}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of
elements (or rows and/or columns) to operate over. If
\code{\link[base]{NULL}}, no subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of
elements (or rows and/or columns) to operate over. If
\code{\link[base]{NULL}}, no subsetting is done.}

\item{...}{Not used.}

\item{force_block_processing}{\code{FALSE} (the default) means that a
seed-aware, optimised method is used (if available). This can be overridden
to use the general block-processing strategy by setting this to \code{TRUE}
(typically not advised).}
}
\value{
Returns \code{\link[base:logical]{TRUE}} if a missing value was
detected, otherwise \code{\link[base:logical]{FALSE}}.
}
\description{
Checks if there are any missing values in an object or not.
\emph{Please use \code{base::anyNA()} instead of \code{anyMissing()},
\code{colAnyNAs()} instead of \code{colAnyMissings()}, and
\code{rowAnyNAs()} instead of \code{rowAnyMissings()}.}
}
\details{
The implementation of this method is optimized for both speed and memory.
The method will return \code{\link[base:logical]{TRUE}} as soon as a missing
value is detected.
}
\seealso{
Starting with R v3.1.0, there is \code{anyNA()} in the \pkg{base},
which provides the same functionality as \code{anyMissing()}.
}
