% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colAvgsPerRowSet.R
\docType{methods}
\name{colAvgsPerRowSet}
\alias{colAvgsPerRowSet}
\alias{colAvgsPerRowSet,DelayedMatrix-method}
\title{Applies a row-by-row (column-by-column) averaging function to equally-sized
subsets of matrix columns (rows)}
\usage{
colAvgsPerRowSet(X, W = NULL, cols = NULL, S, FUN = colMeans,
  tFUN = FALSE, ...)

\S4method{colAvgsPerRowSet}{DelayedMatrix}(X, W = NULL, cols = NULL, S,
  FUN = colMeans, tFUN = FALSE, force_block_processing = FALSE, ...)
}
\arguments{
\item{X}{A \code{\link[base]{numeric}} NxM \code{\link[base]{matrix}}.}

\item{W}{An optional \code{\link[base]{numeric}} NxM
\code{\link[base]{matrix}} of weights.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{S}{An \code{\link[base]{integer}} KxJ \code{\link[base]{matrix}}
specifying the J subsets.  Each column holds K column (row) indices for the
corresponding subset.}

\item{FUN}{The row-by-row (column-by-column) \code{\link[base]{function}}
used to average over each subset of \code{X}.  This function must accept a
\code{\link[base]{numeric}} NxK (KxM) \code{\link[base]{matrix}} and the
\code{\link[base]{logical}} argument \code{na.rm} (which is automatically
set), and return a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (M).}

\item{tFUN}{If \code{\link[base:logical]{TRUE}}, the NxK (KxM)
\code{\link[base]{matrix}} passed to \code{FUN()} is transposed first.}

\item{...}{Additional arguments passed to then \code{FUN}
\code{\link[base]{function}}.}

\item{force_block_processing}{\code{FALSE} (the default) means that a
seed-aware, optimised method is used (if available). This can be overridden
to use the general block-processing strategy by setting this to \code{TRUE}
(typically not advised). The block-processing strategy loads one or more
(depending on \code{getOption("DelayedArray.block.size")}) columns (\code{colFoo()})
or rows (\code{rowFoo()}) into memory as an ordinary \link[base:array]{base::array}.}
}
\value{
Returns a \code{\link[base]{numeric}} JxN (MxJ)
\code{\link[base]{matrix}}, where row names equal \code{rownames(X)}
(\code{colnames(S)}) and column names \code{colnames(S)}
(\code{colnames(X)}).
}
\description{
Applies a row-by-row (column-by-column) averaging function to equally-sized
subsets of matrix columns (rows).  Each subset is averaged independently of
the others.
}
\details{
If argument \code{S} is a single column vector with indices \code{1:N}, then
\code{rowAvgsPerColSet(X, S = S, FUN = rowMeans)} gives the same result as
\code{rowMeans(X)}.  Analogously, for \code{rowAvgsPerColSet()}.
}
