% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colLogSumExps.R,
%   R/rowLogSumExps.R
\docType{methods}
\name{colLogSumExps}
\alias{colLogSumExps}
\alias{rowLogSumExps}
\alias{colLogSumExps,DelayedMatrix-method}
\alias{rowLogSumExps,DelayedMatrix-method}
\title{Accurately computes the logarithm of the sum of exponentials across rows or
columns}
\usage{
colLogSumExps(lx, rows = NULL, cols = NULL, na.rm = FALSE,
  dim. = dim(lx), ...)

rowLogSumExps(lx, rows = NULL, cols = NULL, na.rm = FALSE,
  dim. = dim(lx), ...)

\S4method{colLogSumExps}{DelayedMatrix}(lx, rows = NULL, cols = NULL,
  na.rm = FALSE, dim. = dim(lx), force_block_processing = FALSE, ...)

\S4method{rowLogSumExps}{DelayedMatrix}(lx, rows = NULL, cols = NULL,
  na.rm = FALSE, dim. = dim(lx), force_block_processing = FALSE, ...)
}
\arguments{
\item{lx}{A \code{\link[base]{numeric}} NxK \code{\link[base]{matrix}}.
Typically \code{lx} are \eqn{log(x)} values.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, any missing values are
ignored, otherwise not.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, also when not a
\code{\link[base]{matrix}}.}

\item{...}{Not used.}

\item{force_block_processing}{\code{FALSE} (the default) means that a
seed-aware, optimised method is used (if available). This can be overridden
to use the general block-processing strategy by setting this to \code{TRUE}
(typically not advised).}
}
\value{
A \code{\link[base]{numeric}} \code{\link[base]{vector}} of length N
(K).
}
\description{
Accurately computes the logarithm of the sum of exponentials across rows or
columns.
}
\section{Benchmarking}{


These methods are implemented in native code and have been optimized for
speed and memory.

}

\seealso{
To calculate the same on vectors, \code{\link{logSumExp}}().
}
