% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colMads.R, R/rowMads.R
\docType{methods}
\name{colMads}
\alias{colMads}
\alias{rowMads}
\alias{colMads,DelayedMatrix-method}
\alias{rowMads,DelayedMatrix-method}
\title{Standard deviation estimates for each row (column) in a matrix}
\usage{
colMads(x, rows = NULL, cols = NULL, center = NULL, constant = 1.4826,
  na.rm = FALSE, dim. = dim(x), centers = NULL, ...)

rowMads(x, rows = NULL, cols = NULL, center = NULL, constant = 1.4826,
  na.rm = FALSE, dim. = dim(x), centers = NULL, ...)

\S4method{colMads}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  center = NULL, constant = 1.4826, na.rm = FALSE, dim. = dim(x),
  centers = NULL, force_block_processing = FALSE, ...)

\S4method{rowMads}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  center = NULL, constant = 1.4826, na.rm = FALSE, dim. = dim(x),
  centers = NULL, force_block_processing = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link[base]{numeric}} NxK \code{\link[base]{matrix}}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{center}{A optional \code{\link[base]{numeric}}
\code{\link[base]{vector}} of length N (K) with centers.  By default, they
are calculated using \code{\link{rowMedians}}().}

\item{constant}{A scale factor.  See \code{\link[stats]{mad}} for details.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, missing values are removed
first, otherwise not.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, also when not a
\code{\link[base]{matrix}}.}

\item{centers}{(deprectated) use \code{center} instead.}

\item{...}{Additional arguments passed to \code{\link{rowVars}}() and
\code{\link{rowMedians}}(), respectively.}

\item{force_block_processing}{\code{FALSE} (the default) means that a
seed-aware, optimised method is used (if available). This can be overridden
to use the general block-processing strategy by setting this to \code{TRUE}
(typically not advised). The block-processing strategy loads one or more
(depending on \code{getOption("DelayedArray.block.size")}) columns (\code{colFoo()})
or rows (\code{rowFoo()}) into memory as an ordinary \link[base:array]{base::array}.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Standard deviation estimates for each row (column) in a matrix.
}
\seealso{
\code{\link[stats]{sd}}, \code{\link[stats]{mad}} and
\code{\link[stats:cor]{var}}.  \code{\link{rowIQRs}}().
}
