% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colSds.R, R/rowSds.R
\docType{methods}
\name{colSds}
\alias{colSds}
\alias{rowSds}
\alias{colSds,DelayedMatrix-method}
\alias{rowSds,DelayedMatrix-method}
\title{Standard deviation estimates for each row (column) in a matrix}
\usage{
colSds(x, rows = NULL, cols = NULL, ...)

rowSds(x, rows = NULL, cols = NULL, ...)

\S4method{colSds}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  force_block_processing = FALSE, ...)

\S4method{rowSds}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  force_block_processing = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link[base]{numeric}} NxK \code{\link[base]{matrix}}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{...}{Additional arguments passed to \code{\link{rowVars}}() and
\code{\link{rowMedians}}(), respectively.}

\item{force_block_processing}{\code{FALSE} (the default) means that a
seed-aware, optimised method is used (if available). This can be overridden
to use the general block-processing strategy by setting this to \code{TRUE}
(typically not advised). The block-processing strategy loads one or more
(depending on \code{getOption("DelayedArray.block.size")}) columns (\code{colFoo()})
or rows (\code{rowFoo()}) into memory as an ordinary \link[base:array]{base::array}.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Standard deviation estimates for each row (column) in a matrix.
}
\seealso{
\code{\link[stats]{sd}}, \code{\link[stats]{mad}} and
\code{\link[stats:cor]{var}}.  \code{\link{rowIQRs}}().
}
