% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colSums2.R, R/rowSums2.R
\docType{methods}
\name{colSums2}
\alias{colSums2}
\alias{rowSums2}
\alias{colSums2,DelayedMatrix-method}
\alias{colSums2,Matrix-method}
\alias{colSums2,SolidRleArraySeed-method}
\alias{colSums2,HDF5Matrix-method}
\alias{rowSums2,DelayedMatrix-method}
\alias{rowSums2,Matrix-method}
\title{Calculates the sum for each row (column) in a matrix}
\usage{
colSums2(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x), ...)

rowSums2(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x), ...)

\S4method{colSums2}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, dim. = dim(x), force_block_processing = FALSE, ...)

\S4method{colSums2}{Matrix}(x, rows = NULL, cols = NULL, na.rm = FALSE,
  dim. = dim(x), ...)

\S4method{colSums2}{SolidRleArraySeed}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, dim. = dim(x), ...)

\S4method{colSums2}{HDF5Matrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, dim. = dim(x), ...)

\S4method{rowSums2}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, dim. = dim(x), force_block_processing = FALSE, ...)

\S4method{rowSums2}{Matrix}(x, rows = NULL, cols = NULL, na.rm = FALSE,
  dim. = dim(x), ...)
}
\arguments{
\item{x}{A \code{\link[base]{numeric}} or a \code{\link[base]{logical}}
NxK \code{\link[base]{matrix}}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, also when not a
\code{\link[base]{matrix}}.}

\item{...}{Not used.}

\item{force_block_processing}{\code{FALSE} (the default) means that a
seed-aware, optimised method is used (if available). This can be overridden
to use the general block-processing strategy by setting this to \code{TRUE}
(typically not advised). The block-processing strategy loads one or more
(depending on \code{getOption("DelayedArray.block.size")}) columns (\code{colFoo()})
or rows (\code{rowFoo()}) into memory as an ordinary \link[base:array]{base::array}.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the sum for each row (column) in a matrix.
}
\details{
The implementation of \code{rowSums2()} and \code{colSums2()} is
optimized for both speed and memory.
}
