% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colVars.R, R/rowVars.R
\name{colVars,DelayedMatrix-method}
\alias{colVars,DelayedMatrix-method}
\alias{rowVars,DelayedMatrix-method}
\title{Calculates the variance for each row (column) of a matrix-like object}
\usage{
\S4method{colVars}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, center = NULL, dim. = dim(x),
  force_block_processing = FALSE, ...)

\S4method{rowVars}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, center = NULL, dim. = dim(x),
  force_block_processing = FALSE, ...)
}
\arguments{
\item{x}{A NxK \linkS4class{DelayedMatrix}.}

\item{rows}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{center}{(optional) the center, defaults to the row means.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, essential when x is a
\code{\link[base]{numeric}} vector. Note, that this is not a generic
argument and not all methods need provide it.}

\item{force_block_processing}{\code{FALSE} (the default) means that a
seed-aware, optimised method is used (if available). This can be overridden
to use the general block-processing strategy by setting this to \code{TRUE}
(typically not advised). The block-processing strategy loads one or more
columns (\code{colFoo()}) or rows (\code{rowFoo()}) into memory as an ordinary
\link[base:array]{base::array}, with the number of columns/rows depending on
\link[DelayedArray]{getAutoBlockSize}()).}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the variance for each row (column) of a matrix-like object.
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}},
\code{\link[base]{array}}, or \code{\link[base]{numeric}} call
\code{matrixStats::\link[matrixStats]{rowVars}}
/ \code{matrixStats::\link[matrixStats]{colVars}}.
}
\examples{
# A DelayedMatrix with a 'matrix' seed
dm_matrix <- DelayedArray(matrix(c(rep(1L, 5),
                                   as.integer((0:4) ^ 2),
                                   seq(-5L, -1L, 1L)),
                                 ncol = 3))
# A DelayedMatrix with a 'HDF5ArraySeed' seed
# NOTE: Requires that the HDF5Array package is installed
library(HDF5Array)
dm_HDF5 <- writeHDF5Array(matrix(c(rep(1L, 5),
                                   as.integer((0:4) ^ 2),
                                   seq(-5L, -1L, 1L)),
                                 ncol = 3))

colVars(dm_matrix)

rowVars(dm_matrix)
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowVars}()} and
\code{matrixStats::\link[matrixStats]{colVars}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item For mean estimates, see \code{\link[MatrixGenerics]{rowMeans2}()} and
\code{\link[base:colSums]{rowMeans}()}.
\item For standard deviation estimates, see \code{\link[MatrixGenerics]{rowSds}()}.
\item \code{stats::\link[stats]{var}()}.
}
}
\author{
Peter Hickey
}
