% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{DelayedMatrixStats-defunct}
\alias{DelayedMatrixStats-defunct}
\alias{colAnyMissings}
\alias{rowAnyMissings}
\title{Defunct functions in package \sQuote{DelayedMatrixStats}}
\description{
These functions are defunct and no longer available.
}
\details{
Defunct functions are:
\itemize{
\item{colAnyMissings}
\item{rowAnyMissings}
}
}
