% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{colAnyMissings}
\alias{colAnyMissings}
\alias{rowAnyMissings}
\title{Deprecated functions in \sQuote{DelayedMatrixStats}}
\usage{
colAnyMissings(x, rows = NULL, cols = NULL, ...)

rowAnyMissings(x, rows = NULL, cols = NULL, ...)
}
\arguments{
\item{x}{A NxK \linkS4class{DelayedMatrix}.}

\item{rows}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{...}{Additional arguments passed to specific methods.}
}
\description{
These functions are provided for compatibility with older versions of
\sQuote{DelayedMatrixStats} only, and will be defunct at the next release.
}
\details{
The following functions are deprecated and will be made defunct;
use the replacement indicated below:
\itemize{
\item{\code{colAnyMissings()}: \code{\link{colAnyNAs}()}}
\item{\code{rowAnyMissings()}: \code{\link{rowAnyNAs}()}}
}
}
