% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayedMatrixStats-package.r
\docType{package}
\name{DelayedMatrixStats}
\alias{DelayedMatrixStats}
\alias{DelayedMatrixStats-package}
\title{DelayedMatrixStats: Functions that apply to rows and columns of
\emph{DelayedMatrix} objects.}
\description{
\strong{DelayedMatrixStats} is a port of the
\href{https://CRAN.R-project.org/package=matrixStats}{matrixStats} API to
work with \emph{DelayedMatrix} objects from the
\href{http://bioconductor.org/packages/DelayedArray/}{DelayedArray}
package. High-performing functions operating on rows and columns of
\emph{DelayedMatrix} objects, e.g. \code{colMedians()} / \code{rowMedians()},
\code{colRanks()} / \code{rowRanks()}, and \code{colSds()} / \code{rowSds()}. Functions
optimized per data type and for subsetted calculations such that both memory
usage and processing time is minimized.
}
