% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{rowMaxs}
\alias{rowMins}
\alias{rowRanges}
\alias{colMaxs}
\alias{colMins}
\alias{colRanges}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{DelayedArray}{\code{\link[DelayedArray:DelayedMatrix-stats]{colMaxs}}, \code{\link[DelayedArray:DelayedMatrix-stats]{colMins}}, \code{\link[DelayedArray:DelayedMatrix-stats]{colRanges}}, \code{\link[DelayedArray:DelayedMatrix-stats]{rowMaxs}}, \code{\link[DelayedArray:DelayedMatrix-stats]{rowMins}}, \code{\link[DelayedArray:DelayedMatrix-stats]{rowRanges}}}
}}

