% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayedArray_utils.R
\name{rowblock_APPLY}
\alias{rowblock_APPLY}
\title{Adapted from \code{DelayedArray:::colblock_APPLY()}}
\usage{
rowblock_APPLY(x, APPLY, ..., sink = NULL)
}
\description{
Adapted from \code{DelayedArray:::colblock_APPLY()}
}
\keyword{internal}
