% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/rowMedians.R
\docType{methods}
\name{rowMedians}
\alias{rowMedians}
\alias{rowMedians,DelayedMatrix-method}
\title{Calculates the median for each row (column) in a matrix}
\usage{
rowMedians(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x),
  ...)

\S4method{rowMedians}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, dim. = dim(x), force_block_processing = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link[base]{numeric}} NxK \code{\link[base]{matrix}}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, also when not a
\code{\link[base]{matrix}}.}

\item{...}{Not used.}

\item{force_block_processing}{\code{FALSE} (the default) means that a
seed-aware, optimised method is used (if available). This can be overridden
to use the general block-processing strategy by setting this to \code{TRUE}
(typically not advised).}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the median for each row (column) in a matrix.
}
\details{
The implementation of \code{rowMedians()} and \code{colMedians()} is
optimized for both speed and memory.  To avoid coercing to
\code{\link[base]{double}}s (and hence memory allocation), there is a
special implementation for \code{\link[base]{integer}} matrices.  That is,
if \code{x} is an \code{\link[base]{integer}} \code{\link[base]{matrix}},
then \code{rowMedians(as.double(x))} (\code{rowMedians(as.double(x))}) would
require three times the memory of \code{rowMedians(x)}
(\code{colMedians(x)}), but all this is avoided.
}
\seealso{
See \code{\link{rowMedians}}() and \code{colMedians()} for weighted
medians.  For mean estimates, see \code{rowMeans()} in
\code{\link[base]{colSums}}().
}
