% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayedArray_utils.R
\name{seedClass}
\alias{seedClass}
\title{Get the class of the seed of a DelayedArray}
\usage{
seedClass(x)
}
\arguments{
\item{x}{A \link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray}.}
}
\value{
The class of the seed of the \link[DelayedArray:DelayedArray-class]{DelayedArray::DelayedArray}
object.
}
\description{
Get the class of the seed of a DelayedArray
}
\note{
Can't use \code{class(seed(x))} because \code{seed(x)} will return an error if
the seed is a \link[DelayedArray:DelayedOp-class]{DelayedArray::DelayedOp}.
}
\keyword{internal}
