% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colAlls.R
\docType{methods}
\name{colAlls}
\alias{colAlls}
\alias{colAlls,matrix-method}
\alias{colAlls,DelayedMatrix-method}
\alias{colAlls,ANY-method}
\title{Checks if a value exists / does not exist in each row (column) of a matrix}
\usage{
colAlls(x, rows = NULL, cols = NULL, value = TRUE, na.rm = FALSE,
  dim. = dim(x), ...)

\S4method{colAlls}{matrix}(x, rows = NULL, cols = NULL, value = TRUE,
  na.rm = FALSE, dim. = dim(x), ...)

\S4method{colAlls}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  value = TRUE, na.rm = FALSE, dim. = dim(x), ...)

\S4method{colAlls}{ANY}(x, rows = NULL, cols = NULL, value = TRUE,
  na.rm = FALSE, dim. = dim(x), ...)
}
\arguments{
\item{x}{An NxK \code{\link[base]{matrix}} or an N * K
\code{\link[base]{vector}}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of
elements (or rows and/or columns) to operate over. If
\code{\link[base]{NULL}}, no subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of
elements (or rows and/or columns) to operate over. If
\code{\link[base]{NULL}}, no subsetting is done.}

\item{value}{A value to search for.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, also when not a
\code{\link[base]{matrix}}.}

\item{...}{Not used.}
}
\value{
\code{rowAlls()} (\code{colAlls()}) returns an
\code{\link[base]{logical}} \code{\link[base]{vector}} of length N (K).
Analogously for \code{rowAnys()} (\code{rowAlls()}).
}
\description{
Checks if a value exists / does not exist in each row (column) of a matrix.
}
\details{
These functions takes either a matrix or a vector as input. If a vector,
then argument \code{dim.} must be specified and fulfill \code{prod(dim.) ==
length(x)}.  The result will be identical to the results obtained when
passing \code{matrix(x, nrow = dim.[1L], ncol = dim.[2L])}, but avoids
having to temporarily create/allocate a matrix, if only such is needed
only for these calculations.
}
\section{Logical \code{value}}{


When \code{value} is logical, the result is as if the function is applied
on \code{as.logical(x)}.  More specifically, if \code{x} is numeric, then
all zeros are treates as \code{FALSE}, non-zero values as \code{TRUE},
and all missing values as \code{NA}.

}

\seealso{
rowCounts
}
