% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colCollapse.R
\docType{methods}
\name{colCollapse}
\alias{colCollapse}
\alias{colCollapse,matrix-method}
\alias{colCollapse,DelayedMatrix-method}
\alias{colCollapse,ANY-method}
\title{Extracts one cell per row (column) from a matrix}
\usage{
colCollapse(x, idxs, cols = NULL, dim. = dim(x), ...)

\S4method{colCollapse}{matrix}(x, idxs, cols = NULL, dim. = dim(x), ...)

\S4method{colCollapse}{DelayedMatrix}(x, idxs, cols = NULL, dim. = dim(x),
  ...)

\S4method{colCollapse}{ANY}(x, idxs, cols = NULL, dim. = dim(x), ...)
}
\arguments{
\item{x}{An NxK \code{\link[base]{matrix}}.}

\item{idxs}{An index \code{\link[base]{vector}} of (maximum) length N (K)
specifying the columns (rows) to be extracted.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, also when not a
\code{\link[base]{matrix}}.}

\item{...}{Not used.}
}
\value{
Returns a \code{\link[base]{vector}} of length N (K).
}
\description{
Extracts one cell per row (column) from a matrix.  The implementation is
optimized for memory and speed.
}
\seealso{
\emph{Matrix indexing} to index elements in matrices and arrays,
cf. \code{\link[base]{[}}().
}
