% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colCummaxs.R, R/colCummins.R,
%   R/colCumprods.R, R/colCumsums.R, R/rowCummaxs.R, R/rowCummins.R,
%   R/rowCumprods.R, R/rowCumsums.R
\docType{methods}
\name{colCummaxs}
\alias{colCummaxs}
\alias{colCummins}
\alias{colCumprods}
\alias{colCumsums}
\alias{rowCummaxs}
\alias{rowCummins}
\alias{rowCumprods}
\alias{rowCumsums}
\alias{colCummaxs,DelayedMatrix-method}
\alias{colCummins,DelayedMatrix-method}
\alias{colCumprods,DelayedMatrix-method}
\alias{colCumsums,DelayedMatrix-method}
\alias{rowCummaxs,DelayedMatrix-method}
\alias{rowCummins,DelayedMatrix-method}
\alias{rowCumprods,DelayedMatrix-method}
\alias{rowCumsums,DelayedMatrix-method}
\title{Cumulative sums, products, minima and maxima for each row (column) in a
matrix}
\usage{
colCummaxs(x, rows = NULL, cols = NULL, dim. = dim(x), ...)

colCummins(x, rows = NULL, cols = NULL, dim. = dim(x), ...)

colCumprods(x, rows = NULL, cols = NULL, dim. = dim(x), ...)

colCumsums(x, rows = NULL, cols = NULL, dim. = dim(x), ...)

rowCummaxs(x, rows = NULL, cols = NULL, dim. = dim(x), ...)

rowCummins(x, rows = NULL, cols = NULL, dim. = dim(x), ...)

rowCumprods(x, rows = NULL, cols = NULL, dim. = dim(x), ...)

rowCumsums(x, rows = NULL, cols = NULL, dim. = dim(x), ...)

\S4method{colCummaxs}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  dim. = dim(x), force_block_processing = FALSE, ...)

\S4method{colCummins}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  dim. = dim(x), force_block_processing = FALSE, ...)

\S4method{colCumprods}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  dim. = dim(x), force_block_processing = FALSE, ...)

\S4method{colCumsums}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  dim. = dim(x), force_block_processing = FALSE, ...)

\S4method{rowCummaxs}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  dim. = dim(x), force_block_processing = FALSE, ...)

\S4method{rowCummins}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  dim. = dim(x), force_block_processing = FALSE, ...)

\S4method{rowCumprods}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  dim. = dim(x), force_block_processing = FALSE, ...)

\S4method{rowCumsums}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  dim. = dim(x), force_block_processing = FALSE, ...)
}
\arguments{
\item{x}{A NxK \linkS4class{DelayedMatrix}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of elements
(or rows and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of elements
(or rows and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, also when not a
\code{\link[base]{matrix}}.}

\item{...}{Additional arguments passed to specific methods.}

\item{force_block_processing}{\code{FALSE} (the default) means that a
seed-aware, optimised method is used (if available). This can be overridden
to use the general block-processing strategy by setting this to \code{TRUE}
(typically not advised). The block-processing strategy loads one or more
(depending on \code{\link[DelayedArray]{getAutoBlockSize}()}) columns (\code{colFoo()})
or rows (\code{rowFoo()}) into memory as an ordinary \link[base:array]{base::array}.}
}
\value{
Returns a \code{\link[base]{numeric}} NxK \code{\link[base]{matrix}}
of the same mode as \code{x}, except when \code{x} is of mode
\code{\link[base]{logical}}, then the return type is
\code{\link[base]{integer}}.
}
\description{
Cumulative sums, products, minima and maxima for each row (column) in a
matrix.
}
\examples{
# A DelayedMatrix with a 'matrix' seed
dm_matrix <- DelayedArray(matrix(c(rep(1L, 5),
                                   as.integer((0:4) ^ 2),
                                   seq(-5L, -1L, 1L)),
                                 ncol = 3))
# A DelayedMatrix with a 'Matrix' seed
dm_Matrix <- DelayedArray(Matrix::Matrix(c(rep(1L, 5),
                                           as.integer((0:4) ^ 2),
                                           seq(-5L, -1L, 1L)),
                                         ncol = 3))

colCummaxs(dm_matrix)

colCummins(dm_matrix)

colCumprods(dm_matrix)

colCumsums(dm_matrix)

# Only use rows 2-4
rowCummaxs(dm_Matrix, rows = 2:4)

# Only use rows 2-4
rowCummins(dm_Matrix, rows = 2:4)

# Only use rows 2-4
rowCumprods(dm_Matrix, rows = 2:4)

# Only use rows 2-4
rowCumsums(dm_Matrix, rows = 2:4)
}
\seealso{
See \code{\link[base]{cumsum}}(), \code{\link[base]{cumprod}}(),
\code{\link[base]{cummin}}(), and \code{\link[base]{cummax}}().
}
