% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colCummaxs.R, R/rowCummaxs.R
\docType{methods}
\name{colCummaxs}
\alias{colCummaxs}
\alias{rowCummaxs}
\alias{colCummaxs,DelayedMatrix-method}
\alias{rowCummaxs,DelayedMatrix-method}
\title{Cumulative sums, products, minima and maxima for each row (column) in a
matrix}
\usage{
colCummaxs(x, rows = NULL, cols = NULL, dim. = dim(x), ...)

rowCummaxs(x, rows = NULL, cols = NULL, dim. = dim(x), ...)

\S4method{colCummaxs}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  dim. = dim(x), force_block_processing = FALSE, ...)

\S4method{rowCummaxs}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  dim. = dim(x), force_block_processing = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link[base]{numeric}} NxK \code{\link[base]{matrix}}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of elements
(or rows and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of elements
(or rows and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, also when not a
\code{\link[base]{matrix}}.}

\item{...}{Not used.}

\item{force_block_processing}{\code{FALSE} (the default) means that a
seed-aware, optimised method is used (if available). This can be overridden
to use the general block-processing strategy by setting this to \code{TRUE}
(typically not advised). The block-processing strategy loads one or more
(depending on \code{getOption("DelayedArray.block.size")}) columns (\code{colFoo()})
or rows (\code{rowFoo()}) into memory as an ordinary \link[base:array]{base::array}.}
}
\value{
Returns a \code{\link[base]{numeric}} NxK \code{\link[base]{matrix}}
of the same mode as \code{x}.
}
\description{
Cumulative sums, products, minima and maxima for each row (column) in a
matrix.
}
\seealso{
See \code{\link[base]{cumsum}}(), \code{\link[base]{cumprod}}(),
\code{\link[base]{cummin}}(), and \code{\link[base]{cummax}}().
}
