% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colDiffs.R, R/rowDiffs.R
\docType{methods}
\name{colDiffs}
\alias{colDiffs}
\alias{rowDiffs}
\alias{colDiffs,DelayedMatrix-method}
\alias{rowDiffs,DelayedMatrix-method}
\title{Calculates difference for each row (column) in a matrix}
\usage{
colDiffs(x, rows = NULL, cols = NULL, lag = 1L, differences = 1L,
  dim. = dim(x), ...)

rowDiffs(x, rows = NULL, cols = NULL, lag = 1L, differences = 1L,
  dim. = dim(x), ...)

\S4method{colDiffs}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  lag = 1L, differences = 1L, dim. = dim(x),
  force_block_processing = FALSE, ...)

\S4method{rowDiffs}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  lag = 1L, differences = 1L, dim. = dim(x),
  force_block_processing = FALSE, ...)
}
\arguments{
\item{x}{A NxK \linkS4class{DelayedMatrix}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{lag}{An \code{\link[base]{integer}} specifying the lag.}

\item{differences}{An \code{\link[base]{integer}} specifying the order of
difference.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, also when not a
\code{\link[base]{matrix}}.}

\item{...}{Additional arguments passed to specific methods.}

\item{force_block_processing}{\code{FALSE} (the default) means that a
seed-aware, optimised method is used (if available). This can be overridden
to use the general block-processing strategy by setting this to \code{TRUE}
(typically not advised). The block-processing strategy loads one or more
(depending on \code{\link[DelayedArray]{getAutoBlockSize}()}) columns (\code{colFoo()})
or rows (\code{rowFoo()}) into memory as an ordinary \link[base:array]{base::array}.}
}
\value{
Returns a \code{\link[base]{numeric}} Nx(K-1) or (N-1)xK
\code{\link[base]{matrix}}.
}
\description{
Calculates difference for each row (column) in a matrix.
}
\examples{
# A DelayedMatrix with a 'matrix' seed
dm_matrix <- DelayedArray(matrix(c(rep(1L, 5),
                                   as.integer((0:4) ^ 2),
                                   seq(-5L, -1L, 1L)),
                                 ncol = 3))
# A DelayedMatrix with a 'HDF5ArraySeed' seed
# NOTE: Requires that the HDF5Array package is installed
library(HDF5Array)
dm_HDF5 <- writeHDF5Array(matrix(c(rep(1L, 5),
                                   as.integer((0:4) ^ 2),
                                   seq(-5L, -1L, 1L)),
                                 ncol = 3))

colDiffs(dm_matrix)

rowDiffs(dm_HDF5)
# In reverse column order
rowDiffs(dm_HDF5, cols = seq(ncol(dm_HDF5), 1, -1))
}
\seealso{
See also \code{\link{diff2}}().
}
