% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colIQRs.R, R/rowIQRs.R
\name{colIQRs,DelayedMatrix-method}
\alias{colIQRs,DelayedMatrix-method}
\alias{rowIQRs,DelayedMatrix-method}
\title{Calculates the interquartile range for each row (column) of a matrix-like
object}
\usage{
\S4method{colIQRs}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, force_block_processing = FALSE, ...)

\S4method{rowIQRs}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, force_block_processing = FALSE, ...)
}
\arguments{
\item{x}{A NxK \linkS4class{DelayedMatrix}.}

\item{rows}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{force_block_processing}{\code{FALSE} (the default) means that a
seed-aware, optimised method is used (if available). This can be overridden
to use the general block-processing strategy by setting this to \code{TRUE}
(typically not advised). The block-processing strategy loads one or more
columns (\code{colFoo()}) or rows (\code{rowFoo()}) into memory as an ordinary
\link[base:array]{base::array}, with the number of columns/rows depending on
\link[DelayedArray]{getAutoBlockSize}()).}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the interquartile range for each row (column) of a matrix-like
object.
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}},
\code{\link[base]{array}}, or \code{\link[base]{numeric}} call
\code{matrixStats::\link[matrixStats]{rowIQRs}}
/ \code{matrixStats::\link[matrixStats]{colIQRs}}.
}
\examples{
# A DelayedMatrix with a 'matrix' seed
dm_matrix <- DelayedArray(matrix(c(rep(1L, 5),
                                   as.integer((0:4) ^ 2),
                                   seq(-5L, -1L, 1L)),
                                 ncol = 3))
# A DelayedMatrix with a 'Matrix' seed
dm_Matrix <- DelayedArray(Matrix::Matrix(c(rep(1L, 5),
                                           as.integer((0:4) ^ 2),
                                           seq(-5L, -1L, 1L)),
                                         ncol = 3))

colIQRs(dm_matrix)

# Only using rows 2-4
rowIQRs(dm_matrix, rows = 2:4)
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowIQRs}()} and
\code{matrixStats::\link[matrixStats]{colIQRs}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item For a non-robust analog, see \code{\link[MatrixGenerics]{rowSds}()}. For a more
robust version see \code{\link[MatrixGenerics:rowMads]{rowMads()}}
\item \code{stats::\link[stats]{IQR}()}.
}
}
\author{
Peter Hickey
}
