% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colIQRs.R, R/rowIQRs.R
\docType{methods}
\name{colIQRs}
\alias{colIQRs}
\alias{rowIQRs}
\alias{colIQRs,DelayedMatrix-method}
\alias{rowIQRs,DelayedMatrix-method}
\title{Estimates of the interquartile range for each row (column) in a matrix}
\usage{
colIQRs(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)

rowIQRs(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{colIQRs}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, force_block_processing = FALSE, ...)

\S4method{rowIQRs}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, force_block_processing = FALSE, ...)
}
\arguments{
\item{x}{A NxK \linkS4class{DelayedMatrix}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of
elements (or rows and/or columns) to operate over. If
\code{\link[base]{NULL}}, no subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of
elements (or rows and/or columns) to operate over. If
\code{\link[base]{NULL}}, no subsetting is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, missing values are dropped
first, otherwise not.}

\item{...}{Additional arguments passed to specific methods.}

\item{force_block_processing}{\code{FALSE} (the default) means that a
seed-aware, optimised method is used (if available). This can be overridden
to use the general block-processing strategy by setting this to \code{TRUE}
(typically not advised). The block-processing strategy loads one or more
(depending on \code{\link[DelayedArray]{getAutoBlockSize}()}) columns (\code{colFoo()})
or rows (\code{rowFoo()}) into memory as an ordinary \link[base:array]{base::array}.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Estimates of the interquartile range for each row (column) in a matrix.
}
\section{Missing values}{

 Contrary to \code{\link[stats]{IQR}}, which gives
an error if there are missing values and \code{na.rm = FALSE}, \code{iqr()}
and its corresponding row and column-specific functions return
\code{\link[base]{NA}}_real_.

}

\examples{
# A DelayedMatrix with a 'matrix' seed
dm_matrix <- DelayedArray(matrix(c(rep(1L, 5),
                                   as.integer((0:4) ^ 2),
                                   seq(-5L, -1L, 1L)),
                                 ncol = 3))
# A DelayedMatrix with a 'Matrix' seed
dm_Matrix <- DelayedArray(Matrix::Matrix(c(rep(1L, 5),
                                           as.integer((0:4) ^ 2),
                                           seq(-5L, -1L, 1L)),
                                         ncol = 3))

colIQRs(dm_matrix)

# Only using rows 2-4
rowIQRs(dm_matrix, rows = 2:4)
}
\seealso{
See \code{\link[stats]{IQR}}.  See \code{\link{rowSds}}().
}
