% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colMads.R, R/colSds.R,
%   R/rowMads.R, R/rowSds.R
\docType{methods}
\name{colMads}
\alias{colMads}
\alias{colSds}
\alias{rowMads}
\alias{rowSds}
\alias{colMads,DelayedMatrix-method}
\alias{colSds,DelayedMatrix-method}
\alias{rowMads,DelayedMatrix-method}
\alias{rowSds,DelayedMatrix-method}
\title{Standard deviation estimates for each row (column) in a matrix}
\usage{
colMads(x, rows = NULL, cols = NULL, center = NULL, constant = 1.4826,
  na.rm = FALSE, dim. = dim(x), ...)

colSds(x, rows = NULL, cols = NULL, ...)

rowMads(x, rows = NULL, cols = NULL, center = NULL, constant = 1.4826,
  na.rm = FALSE, dim. = dim(x), ...)

rowSds(x, rows = NULL, cols = NULL, ...)

\S4method{colMads}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  center = NULL, constant = 1.4826, na.rm = FALSE, dim. = dim(x),
  force_block_processing = FALSE, ...)

\S4method{colSds}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  force_block_processing = FALSE, ...)

\S4method{rowMads}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  center = NULL, constant = 1.4826, na.rm = FALSE, dim. = dim(x),
  force_block_processing = FALSE, ...)

\S4method{rowSds}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  force_block_processing = FALSE, ...)
}
\arguments{
\item{x}{A NxK \linkS4class{DelayedMatrix}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{center}{A optional \code{\link[base]{numeric}}
\code{\link[base]{vector}} of length N (K) with centers.  By default, they
are calculated using \code{\link{rowMedians}}().}

\item{constant}{A scale factor.  See \code{\link[stats]{mad}} for details.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, missing values are removed
first, otherwise not.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, also when not a
\code{\link[base]{matrix}}.}

\item{...}{Additional arguments passed to specific methods.}

\item{force_block_processing}{\code{FALSE} (the default) means that a
seed-aware, optimised method is used (if available). This can be overridden
to use the general block-processing strategy by setting this to \code{TRUE}
(typically not advised). The block-processing strategy loads one or more
(depending on \code{getOption("DelayedArray.block.size")}) columns (\code{colFoo()})
or rows (\code{rowFoo()}) into memory as an ordinary \link[base:array]{base::array}.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Standard deviation estimates for each row (column) in a matrix.
}
\examples{
# A DelayedMatrix with a 'data.frame' seed
dm_df <- DelayedArray(data.frame(C1 = rep(1L, 5),
                                 C2 = as.integer((0:4) ^ 2),
                                 C3 = seq(-5L, -1L, 1L)))
# A DelayedMatrix with a 'DataFrame' seed
dm_DF <- DelayedArray(S4Vectors::DataFrame(C1 = rep(1L, 5),
                                           C2 = as.integer((0:4) ^ 2),
                                           C3 = seq(-5L, -1L, 1L)))

colMads(dm_df)

colSds(dm_df)

rowMads(dm_DF)

rowSds(dm_DF)
}
\seealso{
\code{\link[stats]{sd}}, \code{\link[stats]{mad}} and
\code{\link[stats:cor]{var}}.  \code{\link{rowIQRs}}().
}
