% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colMeans2.R, R/rowMeans2.R
\docType{methods}
\name{colMeans2}
\alias{colMeans2}
\alias{rowMeans2}
\alias{colMeans2,DelayedMatrix-method}
\alias{colMeans2,Matrix-method}
\alias{colMeans2,SolidRleArraySeed-method}
\alias{rowMeans2,DelayedMatrix-method}
\alias{rowMeans2,Matrix-method}
\title{Calculates the mean for each row (column) in a matrix}
\usage{
colMeans2(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x),
  ...)

rowMeans2(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x),
  ...)

\S4method{colMeans2}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, dim. = dim(x), force_block_processing = FALSE, ...)

\S4method{colMeans2}{Matrix}(x, rows = NULL, cols = NULL, na.rm = FALSE,
  dim. = dim(x), ...)

\S4method{colMeans2}{SolidRleArraySeed}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, dim. = dim(x), ...)

\S4method{rowMeans2}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, dim. = dim(x), force_block_processing = FALSE, ...)

\S4method{rowMeans2}{Matrix}(x, rows = NULL, cols = NULL, na.rm = FALSE,
  dim. = dim(x), ...)
}
\arguments{
\item{x}{A \code{\link[base]{numeric}} NxK \code{\link[base]{matrix}}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, also when not a
\code{\link[base]{matrix}}.}

\item{...}{Not used.}

\item{force_block_processing}{\code{FALSE} (the default) means that a
seed-aware, optimised method is used (if available). This can be overridden
to use the general block-processing strategy by setting this to \code{TRUE}
(typically not advised).}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the mean for each row (column) in a matrix.
}
\details{
The implementation of \code{rowMeans2()} and \code{colMeans2()} is
optimized for both speed and memory.
}
