% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colOrderStats.R,
%   R/rowOrderStats.R
\docType{methods}
\name{colOrderStats}
\alias{colOrderStats}
\alias{rowOrderStats}
\alias{colOrderStats,DelayedMatrix-method}
\alias{rowOrderStats,DelayedMatrix-method}
\title{Gets an order statistic for each row (column) in a matrix}
\usage{
colOrderStats(x, rows = NULL, cols = NULL, which, dim. = dim(x), ...)

rowOrderStats(x, rows = NULL, cols = NULL, which, dim. = dim(x), ...)

\S4method{colOrderStats}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  which, dim. = dim(x), force_block_processing = FALSE, ...)

\S4method{rowOrderStats}{DelayedMatrix}(x, rows = NULL, cols = NULL,
  which, dim. = dim(x), force_block_processing = FALSE, ...)
}
\arguments{
\item{x}{A NxK \linkS4class{DelayedMatrix}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{which}{An \code{\link[base]{integer}} index in [1,K] ([1,N])
indicating which order statistic to be returned.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, also when not a
\code{\link[base]{matrix}}.}

\item{...}{Additional arguments passed to specific methods.}

\item{force_block_processing}{\code{FALSE} (the default) means that a
seed-aware, optimised method is used (if available). This can be overridden
to use the general block-processing strategy by setting this to \code{TRUE}
(typically not advised). The block-processing strategy loads one or more
(depending on \code{\link[DelayedArray]{getDefaultBlockSize}()}) columns (\code{colFoo()})
or rows (\code{rowFoo()}) into memory as an ordinary \link[base:array]{base::array}.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Gets an order statistic for each row (column) in a matrix.
}
\details{
The implementation of \code{rowOrderStats()} is optimized for both speed and
memory.  To avoid coercing to \code{\link[base]{double}}s (and hence memory
allocation), there is a unique implementation for
\code{\link[base]{integer}} matrices.
}
\section{Missing values}{

 This method does \emph{not} handle missing values,
that is, the result corresponds to having \code{na.rm = FALSE} (if such an
argument would be available).

}

\examples{
# A DelayedMatrix with a 'Matrix' seed
dm_Matrix <- DelayedArray(Matrix::Matrix(c(rep(1L, 5),
                                           as.integer((0:4) ^ 2),
                                           seq(-5L, -1L, 1L)),
                                         ncol = 3))
# Only using columns 2-3
colOrderStats(dm_Matrix, cols = 2:3, which = 1)

# Different algorithms, specified by `which`, may give different results
rowOrderStats(dm_Matrix, which = 1)
rowOrderStats(dm_Matrix, which = 2)
}
\seealso{
See \code{rowMeans()} in \code{\link[base]{colSums}}().
}
