% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colsum.R
\docType{methods}
\name{colsum}
\alias{colsum}
\alias{rowsum}
\alias{colsum,ANY-method}
\title{Give Column and Row Sums of a Matrix-Like Object Based on a Grouping Variable}
\usage{
colsum(x, group, reorder = TRUE, ...)

rowsum(x, group, reorder = TRUE, ...)

\S4method{colsum}{ANY}(x, group, reorder = TRUE, na.rm = FALSE, ...)
}
\arguments{
\item{x}{A matrix-like object. Missing values are allowed. A numeric vector
will be treated as a column vector.}

\item{group}{A vector or factor giving the grouping, with one element per
row of \code{x} for \code{rowsum()} or one element per column of \code{x} for \code{colsum()}.
Missing values will be treated as another group and a warning will be given.}

\item{reorder}{If \code{TRUE}, then the result will be in order of
\code{sort(unique(group))}. If \code{FALSE}, it will be in the order that groups are
encountered.}

\item{...}{Additional arguments passed to specific methods.}

\item{na.rm}{logical (\code{TRUE} or \code{FALSE}). Should \code{NA} (including \code{NaN})
values be discarded?}
}
\value{
A matrix-like object containing the sums. For \code{rowsum()}, there will
be one row per unique value of \code{group}. For \code{colsum()}, there will be one
column per unique value of \code{group}.

See \code{\link[base:rowsum]{base::rowsum()}} for the value returned by the default \code{rowsum()}
method.

Specific methods defined in Bioconductor packages will typically return an
object of the same class as the input object.
}
\description{
Compute column and row sums across rows or columns of a numeric
matrix-like object for each level of a grouping variable.

\strong{NOTE}: This man page is for the \code{colsum()} and \code{rowsum()}
\emph{S4 generic functions} defined in the \strong{DelayedMatrixStats} package. See
\code{\link[base:rowsum]{base::rowsum()}} for the default method of \code{rowsum()} (defined in the
\strong{base}) package. The \code{colsum()} generic is the natural extension of
\code{rowsum()} but has no equivalent in the \strong{base} package.
Bioconductor packages can define specific methods for objects (typically
matrix-like) not supported by the default method.
}
\details{
The default is for \code{rowsum()} (resp. \code{colsum()}) to reorder the rows
(columns) to agree with \code{\link[base:tapply]{base::tapply()}} as in the example below.
Reordering should not add noticeably to the time except when there are very
many distinct values of \code{group} and \code{x} has few columns (rows).

To sum over all the rows (columns) of a matrix (i.e. a single group) use
\code{\link[=colSums]{colSums()}} (\code{\link[=rowSums]{rowSums()}}), which should be even faster. To sum over a
subset of rows and/or columns of a matrix (i.e. a subset of a single group)
use \code{\link[=colSums2]{colSums2()}} (\code{\link[=rowSums2]{rowSums2()}}).
}
\examples{
rowsum
showMethods("rowsum")
selectMethod("rowsum", "ANY") # the default method

}
\seealso{
\itemize{
\item \code{\link[base:rowsum]{base::rowsum()}} for the default \code{rowsum()} method.
\item \code{\link[methods:showMethods]{methods::showMethods()}} for displaying a summary of the methods defined for a given generic function.
\item \code{\link[methods:selectMethod]{methods::selectMethod()}} for getting the definition of a specific method.
}
}
