% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colWeightedMeans.R
\docType{methods}
\name{colWeightedMeans}
\alias{colWeightedMeans}
\alias{colWeightedMeans,DelayedMatrix-method}
\title{Calculates the weighted means for each row (column) in a matrix}
\usage{
colWeightedMeans(x, w = NULL, rows = NULL, cols = NULL, na.rm = FALSE,
  ...)

\S4method{colWeightedMeans}{DelayedMatrix}(x, w = NULL, rows = NULL,
  cols = NULL, na.rm = FALSE, force_block_processing = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link[base]{numeric}} NxK \code{\link[base]{matrix}}.}

\item{w}{A \code{\link[base]{numeric}} \code{\link[base]{vector}} of length
K (N).}

\item{rows}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, missing values are
excluded from the calculation, otherwise not.}

\item{...}{Not used.}

\item{force_block_processing}{\code{FALSE} (the default) means that a
seed-aware, optimised method is used (if available). This can be overridden
to use the general block-processing strategy by setting this to \code{TRUE}
(typically not advised).}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the weighted means for each row (column) in a matrix.
}
\details{
The implementations of these methods are optimized for both speed and
memory.  If no weights are given, the corresponding
\code{rowMeans()}/\code{colMeans()} is used.
}
\seealso{
See \code{rowMeans()} and \code{colMeans()} in
\code{\link[base]{colSums}}() for non-weighted means.  See also
\code{\link[stats]{weighted.mean}}.
}
