% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/colWeightedSds.R,
%   R/colWeightedVars.R, R/rowWeightedSds.R, R/rowWeightedVars.R
\docType{methods}
\name{colWeightedSds}
\alias{colWeightedSds}
\alias{colWeightedVars}
\alias{rowWeightedSds}
\alias{rowWeightedVars}
\alias{colWeightedSds,DelayedMatrix-method}
\alias{colWeightedVars,DelayedMatrix-method}
\alias{rowWeightedSds,DelayedMatrix-method}
\alias{rowWeightedVars,DelayedMatrix-method}
\title{Weighted variance and weighted standard deviation}
\usage{
colWeightedSds(x, w = NULL, rows = NULL, cols = NULL,
  na.rm = FALSE, ...)

colWeightedVars(x, w = NULL, rows = NULL, cols = NULL,
  na.rm = FALSE, ...)

rowWeightedSds(x, w = NULL, rows = NULL, cols = NULL,
  na.rm = FALSE, ...)

rowWeightedVars(x, w = NULL, rows = NULL, cols = NULL,
  na.rm = FALSE, ...)

\S4method{colWeightedSds}{DelayedMatrix}(x, w = NULL, rows = NULL,
  cols = NULL, na.rm = FALSE, force_block_processing = FALSE, ...)

\S4method{colWeightedVars}{DelayedMatrix}(x, w = NULL, rows = NULL,
  cols = NULL, na.rm = FALSE, force_block_processing = FALSE, ...)

\S4method{rowWeightedSds}{DelayedMatrix}(x, w = NULL, rows = NULL,
  cols = NULL, na.rm = FALSE, force_block_processing = FALSE, ...)

\S4method{rowWeightedVars}{DelayedMatrix}(x, w = NULL, rows = NULL,
  cols = NULL, na.rm = FALSE, force_block_processing = FALSE, ...)
}
\arguments{
\item{x}{A NxK \linkS4class{DelayedMatrix}.}

\item{w}{a vector of weights the same length as \code{x} giving the weights
to use for each element of \code{x}. Negative weights are treated as zero
weights. Default value is equal weight to all values.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of
elements (or rows and/or columns) to operate over. If
\code{\link[base]{NULL}}, no subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of
elements (or rows and/or columns) to operate over. If
\code{\link[base]{NULL}}, no subsetting is done.}

\item{na.rm}{a logical value indicating whether \code{\link[base]{NA}}
values in \code{x} should be stripped before the computation proceeds, or
not.  If \code{\link[base]{NA}}, no check at all for \code{\link[base]{NA}}s
is done.  Default value is \code{\link[base]{NA}} (for efficiency).}

\item{...}{Additional arguments passed to specific methods.}

\item{force_block_processing}{\code{FALSE} (the default) means that a
seed-aware, optimised method is used (if available). This can be overridden
to use the general block-processing strategy by setting this to \code{TRUE}
(typically not advised). The block-processing strategy loads one or more
(depending on \code{\link[DelayedArray]{getDefaultBlockSize}()}) columns (\code{colFoo()})
or rows (\code{rowFoo()}) into memory as an ordinary \link[base:array]{base::array}.}
}
\value{
Returns a \code{\link[base]{numeric}} scalar.
}
\description{
Computes a weighted variance / standard deviation of a numeric vector or
across rows or columns of a matrix.
}
\details{
The estimator used here is the same as the one used by the "unbiased"
estimator of the \bold{Hmisc} package. More specifically,
\code{weightedVar(x, w = w) == Hmisc::wtd.var(x, weights = w)},
}
\section{Missing values}{

 Missing values are dropped at the very beginning,
if argument \code{na.rm} is \code{\link[base:logical]{TRUE}}, otherwise not.

}

\examples{
# A DelayedMatrix with a 'SolidRleArraySeed' seed
dm_Rle <- RleArray(Rle(c(rep(1L, 5),
                         as.integer((0:4) ^ 2),
                         seq(-5L, -1L, 1L))),
                   dim = c(5, 3))

colWeightedSds(dm_Rle, w = 1 / rowMeans2(dm_Rle))

# Specifying weights inversely proportional to rowwise means
colWeightedVars(dm_Rle, w = 1 / rowMeans2(dm_Rle))

# Specifying weights inversely proportional to columnwise means
rowWeightedSds(dm_Rle, w = 1 / colMeans2(dm_Rle))

# Specifying weights inversely proportional to columnwise means
rowWeightedVars(dm_Rle, w = 1 / colMeans2(dm_Rle))
}
\seealso{
For the non-weighted variance, see \code{\link[stats]{var}}.
}
