% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayedMatrixStats-package.R
\docType{package}
\name{DelayedMatrixStats}
\alias{DelayedMatrixStats-package}
\alias{DelayedMatrixStats}
\title{DelayedMatrixStats: Functions that apply to rows and columns of
\emph{DelayedMatrix} objects.}
\description{
\strong{DelayedMatrixStats} is a port of the
\href{https://CRAN.R-project.org/package=matrixStats}{\strong{matrixStats}} API to
work with \emph{DelayedMatrix} objects from the
\href{http://bioconductor.org/packages/DelayedArray/}{\strong{DelayedArray}}
package. High-performing functions operating on rows and columns of
\emph{DelayedMatrix} objects, e.g. \code{colMedians()} / \code{rowMedians()},
\code{colRanks()} / \code{rowRanks()}, and \code{colSds()} / \code{rowSds()}. Functions
optimized per data type and for subsetted calculations such that both memory
usage and processing time is minimized.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PeteHaitch/DelayedMatrixStats}
  \item Report bugs at \url{https://github.com/PeteHaitch/DelayedMatrixStats/issues}
}

}
\author{
\strong{Maintainer}: Peter Hickey \email{peter.hickey@gmail.com} (\href{https://orcid.org/0000-0002-8153-6258}{ORCID})

Other contributors:
\itemize{
  \item Hervé Pagès \email{hpages.on.github@gmail.com} [contributor]
  \item Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com} [contributor]
}

}
\keyword{internal}
